requirements.txt
streamlit
openai
pdfplumber
pytesseract
pillow
pymupdf
requests
faiss-cpu
tiktoken
